It is my PET Project Telegram bot to apply for a consultation - minor fixes and improvements

+++ handlers/client.py
@@ -16,9 +16,7 @@ load_dotenv(find_dotenv())
 async def command_start(message: types.Message):
     await bot.send_message(message.from_user.id,
                            '<b>Приветствую! ✨\nЯ Бот для записи к КПТ и АСТ психологу '
-                           'Лене Кабаковой 🤩\n</b>',
-                           parse_mode=ParseMode.HTML)
-    await bot.send_message(message.from_user.id,
+                           'Лене Кабаковой 🤩\n\n</b>'
                            'Здесь вы можете немного познакомиться с Леной, '
                            'как специалистом, задать вопрос или записаться на сессию.\n\n'
                            '🔸 Прямо сейчас\nЛена ведет прием только онлайн\n\n'
@@ -29,19 +27,27 @@ async def command_start(message: types.Message):
 
 
 async def bio(callback: types.CallbackQuery):
-    await callback.message.answer(read_from_file())
-    await callback.message.answer('<b>Вернуться в меню</b>', reply_markup=inline_m_kb,
+    if inline_m_kb or inline_kb:
+        await callback.message.delete()
+    await callback.message.answer(f'{read_from_file()}\n\n'
+                                  f'<b>Вернуться в меню</b>',
+                                  reply_markup=inline_m_kb,
                                   parse_mode=ParseMode.HTML)
 
 
 async def prices(callback: types.CallbackQuery):
-    await callback.message.answer(read_price())
-    await callback.message.answer('<b>Вернуться в меню</b>', reply_markup=inline_m_kb,
+    if inline_m_kb or inline_kb:
+        await callback.message.delete()
+    await callback.message.answer(f'{read_price()}\n\n'
+                                  f'<b>Вернуться в меню</b>',
+                                  reply_markup=inline_m_kb,
                                   parse_mode=ParseMode.HTML)
 
 
 # REQUEST MENU
 async def show_menu(callback: types.CallbackQuery):
+    if inline_m_kb or inline_kb:
+        await callback.message.delete()
     await callback.message.answer('<B>Меню</b>', reply_markup=inline_kb,
                                   parse_mode=ParseMode.HTML)
 
@@ -50,7 +56,8 @@ async def show_menu(callback: types.CallbackQuery):
 
 
 class FSMClient(StatesGroup):
-    """ Входим в режим машины состояний """
+    """ Start the StatesMachine"""
+
     name = State()
     phone = State()
     gmt = State()
@@ -58,183 +65,207 @@ class FSMClient(StatesGroup):
 
 
 async def writing_on_consult(callback: types.CallbackQuery):
+    if inline_m_kb or inline_kb:
+        await callback.message.delete()
     await FSMClient.name.set()
-    await callback.message.reply('Чтобы оставить заявку, ответьте, пожалуйста, на 4 вопроса.'
-                                 '\n⚠️ Если передумали, достаточно написать: отмена')
-    await callback.message.reply('Шаг 1.\n'
-                                 'Давайте немного познакомимся!☺️\n'
-                                 'Как Вас зовут?')
+    await callback.message.answer('Чтобы оставить заявку, ответьте, пожалуйста, на 4 вопроса.'
+                                  '\n⚠️ Если передумали, достаточно написать: отмена\n\n'
+                                  'Шаг 1.\n'
+                                  'Давайте немного познакомимся!☺️\n'
+                                  'Как Вас зовут?')
 
 
 async def cancel_handler(message: types.Message, state: FSMContext):
     """ CANCEL state """
+
+    await message.delete()
+    if message.from_user.id >= 1:
+        await bot.delete_message(message.chat.id, message_id=message.message_id - 1)
     current_state = await state.get_state()
     if current_state is None:
         return
     await state.finish()
-    await message.reply('OK. Ваша заявка отменена. 😫')
+    await message.answer('OK. Ваша заявка отменена. 👍🏻', reply_markup=inline_m_kb)
 
 
 async def load_name(message: types.Message, state: FSMContext):
     """ Catch name and write in table """
+
+    await message.delete()
+    if message.from_user.id >= 1:
+        await bot.delete_message(message.chat.id, message_id=message.message_id - 1)
+
     async with state.proxy() as data:
         data['name'] = message.text
     await FSMClient.next()
-    await message.reply('Шаг 2.\nВведите Ваш номер телефона')
+    await message.answer('Шаг 2.\nВведите Ваш номер телефона')
 
 
 async def load_phone(message: types.Message, state: FSMContext):
     """ Catch a phone number """
+
+    await message.delete()
+    if message.from_user.id >= 1:
+        await bot.delete_message(message.chat.id, message_id=message.message_id - 1)
     async with state.proxy() as data:
         data['phone_n'] = message.text
     await FSMClient.next()
-    await message.reply('Шаг3.\nНапишите город, в котором вы проживаете. Или укажите часовой пояс.')
+    await message.answer('Шаг3.\nНапишите город, в котором вы проживаете. Или укажите часовой пояс.')
 
 
 async def load_gmt(message: types.Message, state: FSMContext):
     """Catch a Timezone """
+
+    await message.delete()
+    if message.from_user.id >= 1:
+        await bot.delete_message(message.chat.id, message_id=message.message_id - 1)
     async with state.proxy() as data:
         data['gmt'] = message.text
     await FSMClient.next()
-    await message.reply('Шаг4.\nМожете коротко описать ваш запрос, с которым хотите обратиться.')
+    await message.answer('Шаг4.\nМожете коротко описать ваш запрос,\n'
+                         ' с которым хотите обратиться.')
 
 
 async def load_comment(message: types.Message, state: FSMContext):
     """ Catch a comment """
+
+    if message.from_user.id >= 1:
+        await bot.delete_message(message.chat.id, message_id=message.message_id - 1)
     async with state.proxy() as data:
         data['comment'] = message.text
     await sql_add_command(state)
     await state.finish()
-    if await message.answer('Спасибо! Ваша заявка принята.'
-                            ' Лена скоро с вами свяжется!😊'):
+    if await message.answer(f'Спасибо! Ваша заявка принята.\n'
+                            ' Лена скоро с вами свяжется!😊\n\n'
+                            '<b>Вернуться в меню:</b>',
+                            reply_markup=inline_m_kb,
+                            parse_mode=ParseMode.HTML
+                            ):
         await bot.send_message(chat_id=os.getenv('ID_NUM'),
                                text='You have a new query! 😊 '
                                     'Please check the database!')
         await message.delete()
-    await message.answer('<b>Вернуться в меню:</b>',
-                         reply_markup=inline_m_kb,
-                         parse_mode=ParseMode.HTML)
 
 
-""" FAQ """
+""" FAQ part """
 
 
 async def faq(callback: types.CallbackQuery):
-    await callback.message.reply('<b>Часто задаваемые вопросы</b>',
-                                 reply_markup=inline_faq_kb,
-                                 parse_mode=ParseMode.HTML)
+    if inline_m_kb or inline_kb:
+        await callback.message.delete()
+    await callback.message.answer('<b>🔶 Часто задаваемые вопросы</b>',
+                                  reply_markup=inline_faq_kb,
+                                  parse_mode=ParseMode.HTML)
 
 
 async def first_query(callback: types.CallbackQuery):
-    await callback.message.answer('<b>Как понять, что мне нужно к психологу?</b>',
-                                  parse_mode=ParseMode.HTML)
-    await callback.message.answer(arr_j[0])
-    await callback.message.answer('<b>Еще вопросы:</b>',
-                                  reply_markup=inline_faq_kb,
-                                  parse_mode=ParseMode.HTML)
-    await callback.message.answer('<b>Вернуться в меню</b>',
-                                  reply_markup=inline_m_kb,
-                                  parse_mode=ParseMode.HTML)
+    if inline_m_kb or inline_kb:
+        await callback.message.delete()
+    await callback.message.answer(f'<b>🔶 Как понять, что мне нужно к психологу?</b>\n\n'
+                                  f'{arr_j[0]}\n\n'
+                                  f'<b>Еще вопросы:</b>',
+                                  parse_mode=ParseMode.HTML,
+                                  reply_markup=inline_faq_kb)
 
 
 async def second_query(callback: types.CallbackQuery):
-    await callback.message.answer('<b>Как выбрать психолога?👍🏻</b>',
-                                  parse_mode=ParseMode.HTML)
-    await callback.message.answer(arr_j[1])
-    await callback.message.answer('<b>Еще вопросы:</b>',
-                                  reply_markup=inline_faq_kb,
-                                  parse_mode=ParseMode.HTML)
-    await callback.message.answer('<b>Вернуться в меню</b>',
-                                  reply_markup=inline_m_kb,
-                                  parse_mode=ParseMode.HTML)
+    if inline_m_kb or inline_kb:
+        await callback.message.delete()
+    await callback.message.answer(f'<b>🔶 Как выбрать психолога?👍🏻</b>\n\n'
+                                  f'{arr_j[1]}\n\n'
+                                  f'<b>Еще вопросы:</b>',
+                                  parse_mode=ParseMode.HTML,
+                                  reply_markup=inline_faq_kb)
 
 
 async def third_query(callback: types.CallbackQuery):
-    await callback.message.answer('<b>С какими запросами я могу обратиться?</b>',
-                                  parse_mode=ParseMode.HTML)
-    await callback.message.answer(arr_j[2])
-    await callback.message.answer('<b>Еще вопросы:</b>',
-                                  reply_markup=inline_faq_kb,
-                                  parse_mode=ParseMode.HTML)
-    await callback.message.answer('<b>Вернуться в меню</b>',
-                                  reply_markup=inline_m_kb,
-                                  parse_mode=ParseMode.HTML)
+    if inline_m_kb or inline_kb:
+        await callback.message.delete()
+    await callback.message.answer(f'<b>🔶 С какими запросами я могу обратиться?</b>\n\n'
+                                  f'{arr_j[2]}\n\n'
+                                  f'<b>Еще вопросы:</b>',
+                                  parse_mode=ParseMode.HTML,
+                                  reply_markup=inline_faq_kb)
 
 
 async def four_query(callback: types.CallbackQuery):
-    await callback.message.answer('<b>Как проходит сессия?</b>', parse_mode=ParseMode.HTML)
-    await callback.message.answer(arr_j[3])
-    await callback.message.answer('<b>Еще вопросы:</b>',
-                                  reply_markup=inline_faq_kb,
-                                  parse_mode=ParseMode.HTML)
-    await callback.message.answer('<b>Вернуться в меню</b>',
-                                  reply_markup=inline_m_kb,
-                                  parse_mode=ParseMode.HTML)
+    if inline_m_kb or inline_kb:
+        await callback.message.delete()
+    await callback.message.answer(f'<b>🔶 Как проходит сессия?</b>\n\n'
+                                  f'{arr_j[3]}\n\n'
+                                  f'<b>Еще вопросы:</b>',
+                                  parse_mode=ParseMode.HTML,
+                                  reply_markup=inline_faq_kb)
 
 
 async def five_query(callback: types.CallbackQuery):
-    await callback.message.answer('<b>Как часто нужно встречаться?</b>',
-                                  parse_mode=ParseMode.HTML)
-    await callback.message.answer(arr_j[4])
-    await callback.message.answer('<b>Еще вопросы:</b>',
-                                  reply_markup=inline_faq_kb,
-                                  parse_mode=ParseMode.HTML)
-    await callback.message.answer('<b>Вернуться в меню</b>',
-                                  reply_markup=inline_m_kb,
-                                  parse_mode=ParseMode.HTML)
+    if inline_m_kb or inline_kb:
+        await callback.message.delete()
+    await callback.message.answer(f'<b>🔶 Как часто нужно встречаться?</b>\n\n'
+                                  f'{arr_j[4]}\n\n'
+                                  f'<b>Еще вопросы:</b>',
+                                  parse_mode=ParseMode.HTML,
+                                  reply_markup=inline_faq_kb)
 
 
 async def six_query(callback: types.CallbackQuery):
-    await callback.message.answer('<b>Сколько нужно сессий, для результата?</b>',
-                                  parse_mode=ParseMode.HTML)
-    await callback.message.answer(arr_j[5])
-    await callback.message.answer('<b>Еще вопросы:</b>',
-                                  reply_markup=inline_faq_kb,
-                                  parse_mode=ParseMode.HTML)
-    await callback.message.answer('<b>Вернуться в меню</b>',
-                                  reply_markup=inline_m_kb,
-                                  parse_mode=ParseMode.HTML)
+    if inline_m_kb or inline_kb:
+        await callback.message.delete()
+    await callback.message.answer(f'<b>🔶 Сколько нужно сессий, для результата?</b>\n\n'
+                                  f'{arr_j[5]}\n\n'
+                                  f'<b>Еще вопросы:</b>',
+                                  parse_mode=ParseMode.HTML,
+                                  reply_markup=inline_faq_kb)
 
 
 async def seven_query(callback: types.CallbackQuery):
-    await callback.message.answer('<b>Психолог, психотерапевт и психиатр - в чем отличия?</b>',
-                                  parse_mode=ParseMode.HTML)
-    await callback.message.answer(arr_j[6])
-    await callback.message.answer('<b>Еще вопросы:</b>',
-                                  reply_markup=inline_faq_kb,
-                                  parse_mode=ParseMode.HTML)
-    await callback.message.answer('<b>Вернуться в меню</b>',
-                                  reply_markup=inline_m_kb,
-                                  parse_mode=ParseMode.HTML)
+    if inline_m_kb or inline_kb:
+        await callback.message.delete()
+    await callback.message.answer(f'<b>🔶 Психолог, психотерапевт и психиатр - в чем отличия?</b>\n\n'
+                                  f'{arr_j[6]}\n\n'
+                                  f'<b>Еще вопросы:</b>',
+                                  parse_mode=ParseMode.HTML,
+                                  reply_markup=inline_faq_kb)
+
+
+"""Send Your Question part"""
 
 
 class FormQuestion(StatesGroup):
-    """ Register states """
+    """ Start the StatesMachine """
     question = State()
     name = State()
     phone_n = State()
 
 
 async def send_question(callback: types.CallbackQuery):
+    if inline_m_kb or inline_kb:
+        await callback.message.delete()
     await FormQuestion.question.set()
-    await callback.message.answer('<b>Напишите свой вопрос</b>\n'
+    await callback.message.answer('<b>🔶 Напишите свой вопрос</b>\n'
                                   ' и Лена ответит вам в ближайшее время.\n'
                                   'или напишите: <b>отмена</b>', parse_mode=ParseMode.HTML)
 
 
-# @dp.message_handler(Text(equals='отмена', ignore_case=True), state="*")
 async def cancel_state(message: types.Message, state: FSMContext):
-    """ Cancel states"""
+    """ Cancel states """
+
+    await message.delete()
+    if message.from_user.id >= 1:
+        await bot.delete_message(message.chat.id, message_id=message.message_id - 1)
     cur_state = await state.get_state()
     if cur_state is None:
         return
     await state.finish()
-    await message.reply('OK. Отменяю. 😫')
+    await message.reply('OK. 👍🏻', reply_markup=inline_m_kb)
 
 
-# @dp.message_handler(state=FormQuestion.question)
 async def load_question(message: types.Message, state: FSMContext):
-    """ catch a first state"""
+    """ Catch a first state """
+
+    await message.delete()
+    if message.from_user.id >= 1:
+        await bot.delete_message(message.chat.id, message_id=message.message_id - 1)
     async with state.proxy() as data:
         data['question'] = message.text
 
@@ -242,44 +273,54 @@ async def load_question(message: types.Message, state: FSMContext):
     await message.answer('Напишите свое имя:')
 
 
-# @dp.message_handler(state=FormQuestion.name)
 async def load_name2(message: types.Message, state: FSMContext):
     """ Catch a second state """
+
+    await message.delete()
+    if message.from_user.id >= 1:
+        await bot.delete_message(message.chat.id, message_id=message.message_id - 1)
     async with state.proxy() as data:
         data['name'] = message.text
 
     await FormQuestion.next()
-    await message.answer('Укажите номер телефона,'
+    await message.answer('Укажите ваш номер телефона,'
                          '\nчтобы Лена вам ответила:'
                          )
 
 
-# @dp.message_handler(state=FormQuestion.phone_n)
 async def load_phone_n(message: types.Message, state: FSMContext):
-    """ Catch a third state and send it all in db"""
+    """ Catch a third state and send it all in db """
+    if message.from_user.id >= 1:
+        await bot.delete_message(message.chat.id, message_id=message.message_id - 1)
     async with state.proxy() as data:
         data['phone_n'] = message.text
     await sql_add_command2(state)
     await state.finish()
     if await message.answer('Спасибо!\n Ваш вопрос принят.\n'
-                            ' Лена скоро с вами свяжется!😊'):
+                            ' Лена скоро с вами свяжется!😊\n\n'
+                            '<b>Вернуться в меню:</b>',
+                            reply_markup=inline_m_kb,
+                            parse_mode=ParseMode.HTML):
         await bot.send_message(chat_id=os.getenv('ID_NUM'),
                                text='Вам задали вопрос! 😊 '
                                     'Please check the database with Questions!')
         await message.delete()
-    await message.answer('<b>Вернуться в меню:</b>',
-                         reply_markup=inline_m_kb,
-                         parse_mode=ParseMode.HTML)
+
+
+""" Register handlers part """
 
 
 def register_handlers_client(dp: Dispatcher):
+
     """ Menu buttons handlers """
+
     dp.register_message_handler(command_start, commands=['start', 'help'])
     dp.register_callback_query_handler(bio, Text(startswith='/bio'))
     dp.register_callback_query_handler(prices, Text(startswith='/price'))
     dp.register_callback_query_handler(show_menu, Text(startswith='/Menu'))
 
-    ''' register handlers for creating values for writing in SQLite table '''
+    """ Register handlers for creating values for writing in SQLite table """
+
     dp.register_callback_query_handler(writing_on_consult, Text(startswith='/write'), state=None)
     dp.register_message_handler(cancel_handler, Text(equals='отмена', ignore_case=True), state="*")
     dp.register_message_handler(load_name, state=FSMClient.name)
@@ -287,7 +328,8 @@ def register_handlers_client(dp: Dispatcher):
     dp.register_message_handler(load_gmt, state=FSMClient.gmt)
     dp.register_message_handler(load_comment, state=FSMClient.comment)
 
-    ''' FAQ reg_handlers '''
+    """ FAQ reg_handlers """
+
     dp.register_callback_query_handler(faq, Text(startswith='/faq'))
     dp.register_callback_query_handler(first_query, Text(startswith='/first_query'))
     dp.register_callback_query_handler(second_query, Text(startswith='/second_query'))
@@ -296,7 +338,9 @@ def register_handlers_client(dp: Dispatcher):
     dp.register_callback_query_handler(five_query, Text(startswith='/five_query'))
     dp.register_callback_query_handler(six_query, Text(startswith='/six_query'))
     dp.register_callback_query_handler(seven_query, Text(startswith='/seven_query'))
-    '''FAQ send a question'''
+
+    '''FAQ send a question handlers'''
+
     dp.register_callback_query_handler(send_question, Text(startswith='/question'))
     dp.register_message_handler(cancel_state, Text(equals='отмена', ignore_case=True), state="*")
     dp.register_message_handler(load_question, state=FormQuestion.question)
